+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v5.0.2
 
 * made to support Cassandra 5.0.2
+* intra-cassandra span transfers will be done using binary codec
 
 ## v4.1.1